Refactor(service): 조회 관련 부분 서비스 분리
