+++ src/main/java/com/prgms/allen/dining/domain/reservation/ReservationService.java
@@ -9,9 +9,6 @@ import java.util.Map;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -20,10 +17,6 @@ import com.prgms.allen.dining.domain.member.entity.Member;
 import com.prgms.allen.dining.domain.reservation.dto.ReservationAvailableTimesReq;
 import com.prgms.allen.dining.domain.reservation.dto.ReservationAvailableTimesRes;
 import com.prgms.allen.dining.domain.reservation.dto.ReservationCreateReq;
-import com.prgms.allen.dining.domain.reservation.dto.ReservationDetailRes;
-import com.prgms.allen.dining.domain.reservation.dto.ReservationSimpleResForCustomer;
-import com.prgms.allen.dining.domain.reservation.dto.ReservationSimpleResForOwner;
-import com.prgms.allen.dining.domain.reservation.dto.VisitStatus;
 import com.prgms.allen.dining.domain.reservation.dto.VisitorCountPerVisitTimeProj;
 import com.prgms.allen.dining.domain.reservation.entity.Reservation;
 import com.prgms.allen.dining.domain.reservation.entity.ReservationCustomerInput;
@@ -52,23 +45,6 @@ public class ReservationService {
 		this.reservationRepository = reservationRepository;
 		this.restaurantService = restaurantService;
 		this.memberService = memberService;
-	}
-
-	// TODO: Owner 정보 추가하기
-	public Page<ReservationSimpleResForOwner> getRestaurantReservations(
-		//ownerId
-		long restaurantId,
-		ReservationStatus status,
-		Pageable pageable
-	) {
-		final Restaurant restaurant = restaurantService.findById(restaurantId);
-
-		return new PageImpl<>(
-			reservationRepository.findAllByRestaurantAndStatus(restaurant, status, pageable)
-				.stream()
-				.map(ReservationSimpleResForOwner::new)
-				.toList()
-		);
 	}
 
 	@Transactional
@@ -127,31 +103,6 @@ public class ReservationService {
 				)
 			);
 		}
-	}
-
-	public Page<ReservationSimpleResForCustomer> getRestaurantReservations(
-		long customerId,
-		VisitStatus status,
-		Pageable pageable
-	) {
-		final List<ReservationStatus> statuses = status.getStatuses();
-
-		final Member customer = memberService.findCustomerById(customerId);
-
-		return new PageImpl<>(reservationRepository.findAllByCustomerAndStatusIn(customer, statuses, pageable)
-			.stream()
-			.map(ReservationSimpleResForCustomer::new)
-			.toList());
-	}
-
-	public ReservationDetailRes getReservationDetail(Long reservationId, Long customerId) {
-
-		final Member customer = memberService.findCustomerById(customerId);
-
-		return new ReservationDetailRes(reservationRepository.findByIdAndCustomer(reservationId, customer)
-			.orElseThrow(() -> new NotFoundResourceException(
-				MessageFormat.format("Cannot find Reservation entity for reservationId = {0}", reservationId)
-			)));
 	}
 
 	public Reservation findById(Long id) {
