+++ src/main/java/com/prgms/allen/dining/web/domain/owner/reservation/OwnerReservationApi.java
@@ -13,10 +13,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.prgms.allen.dining.domain.reservation.ReservationFindService;
 import com.prgms.allen.dining.domain.reservation.ReservationService;
 import com.prgms.allen.dining.domain.reservation.ReservationStatusUpdateService;
-import com.prgms.allen.dining.domain.reservation.dto.ReservationStatusUpdateReq;
 import com.prgms.allen.dining.domain.reservation.dto.ReservationSimpleResForOwner;
+import com.prgms.allen.dining.domain.reservation.dto.ReservationStatusUpdateReq;
 import com.prgms.allen.dining.domain.reservation.entity.ReservationStatus;
 
 @RestController
@@ -25,13 +26,15 @@ public class OwnerReservationApi {
 
 	private final ReservationService reservationService;
 	private final ReservationStatusUpdateService reservationStatusUpdateService;
+	private final ReservationFindService reservationFindService;
 
 	public OwnerReservationApi(
 		ReservationService reservationService,
-		ReservationStatusUpdateService reservationStatusUpdateService
-	) {
+		ReservationStatusUpdateService reservationStatusUpdateService,
+		ReservationFindService reservationFindService) {
 		this.reservationService = reservationService;
 		this.reservationStatusUpdateService = reservationStatusUpdateService;
+		this.reservationFindService = reservationFindService;
 	}
 
 	@GetMapping
@@ -40,7 +43,7 @@ public class OwnerReservationApi {
 		@RequestParam Long restaurantId,
 		Pageable pageable
 	) {
-		return ResponseEntity.ok(reservationService.getRestaurantReservations(
+		return ResponseEntity.ok(reservationFindService.getReservations(
 			restaurantId,
 			reservationStatus,
 			pageable
